[6.x] The `hasAttributes` trait is divided into two: `hasAttributes` and `hasCasts`
