+++ src/Illuminate/Database/Eloquent/Model.php
@@ -20,6 +20,7 @@ use JsonSerializable;
 abstract class Model implements Arrayable, ArrayAccess, Jsonable, JsonSerializable, QueueableEntity, UrlRoutable
 {
     use Concerns\HasAttributes,
+        Concerns\HasCasts,
         Concerns\HasEvents,
         Concerns\HasGlobalScopes,
         Concerns\HasRelationships,
